treating optimization of wide-spectrum and high-resolution phase gratings
This work theoretically investigates wide-spectrum and high-resolution diffraction optical elements (DOE)
that are made of stacks of low-resolution binary phase gratings. In which the two-dimensional grids
in different grating layers are arranged with specified displacements. The common Kinoform algorithm is
remodelled for this multi-scale architecture. Numerical computations show that, with increasing the number
of grating layers, the resolution of diffraction fields can be improved and the stability of optical elements
within broad spectrums is significantly enhanced.